requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.3
chromedriver-autoinstaller==0.6.4
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.6
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.5
django-allauth==65.11.1
django-environ==0.12.0
executing==2.2.1
filelock==3.19.1
fonttools==4.60.1
freetype-py==2.5.1
git-filter-repo==2.47.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
idna==3.10
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mercadopago==2.3.0
nest-asyncio==1.6.0
numpy==2.3.3
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pandasql==0.7.3
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycairo==1.29.0
pycparser==2.22
Pygments==2.19.2
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==6.4.0
PySocks==1.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
reportlab==4.4.5
requests==2.32.5
rlPyCairo==0.4.0
seaborn==0.13.2
selenium==4.37.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
svglib==1.6.0
tinycss2==1.5.1
tornado==6.5.2
traitlets==5.14.3
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
virtualenv==20.34.0
wcwidth==0.2.14
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
whitenoise==6.11.0
wsproto==1.2.0
xhtml2pdf==0.2.17
zxing-cpp==2.3.0
